# Faster gather