@TristanBilot 2024-02-04
Wonderful @awni, impressive speedup for such small modification!